Type Alias esp32h2::dma::out_conf0_ch::W

pub type W = W<OUT_CONF0_CH_SPEC>;

Register OUT_CONF0_CH%s writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn out_rst(&mut self) -> OUT_RST_W<'_, OUT_CONF0_CH_SPEC>

Bit 0 - This bit is used to reset DMA channel 1 Tx FSM and Tx FIFO pointer.

pub fn out_loop_test(&mut self) -> OUT_LOOP_TEST_W<'_, OUT_CONF0_CH_SPEC>

Bit 1 - reserved

pub fn out_auto_wrback(&mut self) -> OUT_AUTO_WRBACK_W<'_, OUT_CONF0_CH_SPEC>

Bit 2 - Set this bit to enable automatic outlink-writeback when all the data in tx buffer has been transmitted.

pub fn out_eof_mode(&mut self) -> OUT_EOF_MODE_W<'_, OUT_CONF0_CH_SPEC>

Bit 3 - EOF flag generation mode when transmitting data. 1: EOF flag for Tx channel 1 is generated when data need to transmit has been popped from FIFO in DMA

pub fn outdscr_burst_en(&mut self) -> OUTDSCR_BURST_EN_W<'_, OUT_CONF0_CH_SPEC>

Bit 4 - Set this bit to 1 to enable INCR burst transfer for Tx channel 1 reading link descriptor when accessing internal SRAM.

pub fn out_data_burst_en( &mut self ) -> OUT_DATA_BURST_EN_W<'_, OUT_CONF0_CH_SPEC>

Bit 5 - Set this bit to 1 to enable INCR burst transfer for Tx channel 1 transmitting data when accessing internal SRAM.

pub fn out_etm_en(&mut self) -> OUT_ETM_EN_W<'_, OUT_CONF0_CH_SPEC>

Bit 6 - Set this bit to 1 to enable etm control mode, dma Tx channel 1 is triggered by etm task.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual