Type Alias esp32h2::efuse::rd_repeat_data0::R

pub type R = R<RD_REPEAT_DATA0_SPEC>;

Register RD_REPEAT_DATA0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rd_dis(&self) -> RD_DIS_R

Bits 0:6 - Represents whether reading of individual eFuse block(block4~block10) is disabled or enabled. 1: disabled. 0: enabled.

pub fn rpt4_reserved0_4(&self) -> RPT4_RESERVED0_4_R

Bit 7 - Reserved.

pub fn dis_icache(&self) -> DIS_ICACHE_R

Bit 8 - Represents whether icache is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_usb_jtag(&self) -> DIS_USB_JTAG_R

Bit 9 - Represents whether the function of usb switch to jtag is disabled or enabled. 1: disabled. 0: enabled.

pub fn powerglitch_en(&self) -> POWERGLITCH_EN_R

Bit 10 - Represents whether power glitch function is enabled. 1: enabled. 0: disabled.

pub fn dis_usb_serial_jtag(&self) -> DIS_USB_SERIAL_JTAG_R

Bit 11 - Represents whether USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_force_download(&self) -> DIS_FORCE_DOWNLOAD_R

Bit 12 - Represents whether the function that forces chip into download mode is disabled or enabled. 1: disabled. 0: enabled.

pub fn spi_download_mspi_dis(&self) -> SPI_DOWNLOAD_MSPI_DIS_R

Bit 13 - Represents whether SPI0 controller during boot_mode_download is disabled or enabled. 1: disabled. 0: enabled.

pub fn dis_can(&self) -> DIS_CAN_R

Bit 14 - Represents whether TWAI function is disabled or enabled. 1: disabled. 0: enabled.

pub fn jtag_sel_enable(&self) -> JTAG_SEL_ENABLE_R

Bit 15 - Represents whether the selection between usb_to_jtag and pad_to_jtag through strapping gpio15 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are equal to 0 is enabled or disabled. 1: enabled. 0: disabled.

pub fn soft_dis_jtag(&self) -> SOFT_DIS_JTAG_R

Bits 16:18 - Represents whether JTAG is disabled in soft way. Odd number: disabled. Even number: enabled.

pub fn dis_pad_jtag(&self) -> DIS_PAD_JTAG_R

Bit 19 - Represents whether JTAG is disabled in the hard way(permanently). 1: disabled. 0: enabled.

pub fn dis_download_manual_encrypt(&self) -> DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Bit 20 - Represents whether flash encrypt function is disabled or enabled(except in SPI boot mode). 1: disabled. 0: enabled.

pub fn usb_drefh(&self) -> USB_DREFH_R

Bits 21:22 - Represents the single-end input threhold vrefh, 1.76 V to 2 V with step of 80 mV.

pub fn usb_drefl(&self) -> USB_DREFL_R

Bits 23:24 - Represents the single-end input threhold vrefl, 1.76 V to 2 V with step of 80 mV.

pub fn usb_exchg_pins(&self) -> USB_EXCHG_PINS_R

Bit 25 - Represents whether the D+ and D- pins is exchanged. 1: exchanged. 0: not exchanged.

pub fn vdd_spi_as_gpio(&self) -> VDD_SPI_AS_GPIO_R

Bit 26 - Represents whether vdd spi pin is functioned as gpio. 1: functioned. 0: not functioned.

pub fn rpt4_reserved0_2(&self) -> RPT4_RESERVED0_2_R

Bits 27:28 - Reserved.

pub fn rpt4_reserved0_1(&self) -> RPT4_RESERVED0_1_R

Bit 29 - Reserved.

pub fn rpt4_reserved0_0(&self) -> RPT4_RESERVED0_0_R

Bits 30:31 - Reserved.