Module esp32h2::ledc::ch_conf0

Configuration register 0 for channel %s

Structs

CH_CONF0_SPEC

Configuration register 0 for channel %s

Type Aliases

IDLE_LV_R

Field IDLE_LV reader - This bit is used to control the output value when channel %s is inactive (when LEDC_SIG_OUT_EN_CH%s is 0).

IDLE_LV_W

Field IDLE_LV writer - This bit is used to control the output value when channel %s is inactive (when LEDC_SIG_OUT_EN_CH%s is 0).

OVF_CNT_EN_R

Field OVF_CNT_EN reader - This bit is used to enable the ovf_cnt of channel %s.

OVF_CNT_EN_W

Field OVF_CNT_EN writer - This bit is used to enable the ovf_cnt of channel %s.

OVF_CNT_RESET_W

Field OVF_CNT_RESET writer - Set this bit to reset the ovf_cnt of channel %s.

OVF_NUM_R

Field OVF_NUM reader - This register is used to configure the maximum times of overflow minus 1. The LEDC_OVF_CNT_CH%s_INT interrupt will be triggered when channel %s overflows for (LEDC_OVF_NUM_CH%s + 1) times.

OVF_NUM_W

Field OVF_NUM writer - This register is used to configure the maximum times of overflow minus 1. The LEDC_OVF_CNT_CH%s_INT interrupt will be triggered when channel %s overflows for (LEDC_OVF_NUM_CH%s + 1) times.

PARA_UP_W

Field PARA_UP writer - This bit is used to update LEDC_HPOINT_CH%s, LEDC_DUTY_START_CH%s, LEDC_SIG_OUT_EN_CH%s, LEDC_TIMER_SEL_CH%s, LEDC_DUTY_NUM_CH%s, LEDC_DUTY_CYCLE_CH%s, LEDC_DUTY_SCALE_CH%s, LEDC_DUTY_INC_CH%s, and LEDC_OVF_CNT_EN_CH%s fields for channel %s, and will be automatically cleared by hardware.

R

Register CH%s_CONF0 reader

SIG_OUT_EN_R

Field SIG_OUT_EN reader - Set this bit to enable signal output on channel %s.

SIG_OUT_EN_W

Field SIG_OUT_EN writer - Set this bit to enable signal output on channel %s.

TIMER_SEL_R

Field TIMER_SEL reader - This field is used to select one of timers for channel %s. 0: select timer0, 1: select timer1, 2: select timer2, 3: select timer3

TIMER_SEL_W

Field TIMER_SEL writer - This field is used to select one of timers for channel %s. 0: select timer0, 1: select timer1, 2: select timer2, 3: select timer3

W

Register CH%s_CONF0 writer