Type Alias esp32h2::mcpwm0::evt_en::W

pub type W = W<EVT_EN_SPEC>;

Register EVT_EN writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn evt_timer0_stop_en(&mut self) -> EVT_TIMER0_STOP_EN_W<'_, EVT_EN_SPEC>

Bit 0 - set this bit high to enable timer0 stop event generate

pub fn evt_timer1_stop_en(&mut self) -> EVT_TIMER1_STOP_EN_W<'_, EVT_EN_SPEC>

Bit 1 - set this bit high to enable timer1 stop event generate

pub fn evt_timer2_stop_en(&mut self) -> EVT_TIMER2_STOP_EN_W<'_, EVT_EN_SPEC>

Bit 2 - set this bit high to enable timer2 stop event generate

pub fn evt_timer0_tez_en(&mut self) -> EVT_TIMER0_TEZ_EN_W<'_, EVT_EN_SPEC>

Bit 3 - set this bit high to enable timer0 equal zero event generate

pub fn evt_timer1_tez_en(&mut self) -> EVT_TIMER1_TEZ_EN_W<'_, EVT_EN_SPEC>

Bit 4 - set this bit high to enable timer1 equal zero event generate

pub fn evt_timer2_tez_en(&mut self) -> EVT_TIMER2_TEZ_EN_W<'_, EVT_EN_SPEC>

Bit 5 - set this bit high to enable timer2 equal zero event generate

pub fn evt_timer0_tep_en(&mut self) -> EVT_TIMER0_TEP_EN_W<'_, EVT_EN_SPEC>

Bit 6 - set this bit high to enable timer0 equal period event generate

pub fn evt_timer1_tep_en(&mut self) -> EVT_TIMER1_TEP_EN_W<'_, EVT_EN_SPEC>

Bit 7 - set this bit high to enable timer1 equal period event generate

pub fn evt_timer2_tep_en(&mut self) -> EVT_TIMER2_TEP_EN_W<'_, EVT_EN_SPEC>

Bit 8 - set this bit high to enable timer2 equal period event generate

pub fn evt_op0_tea_en(&mut self) -> EVT_OP0_TEA_EN_W<'_, EVT_EN_SPEC>

Bit 9 - set this bit high to enable PWM generator0 timer equal a event generate

pub fn evt_op1_tea_en(&mut self) -> EVT_OP1_TEA_EN_W<'_, EVT_EN_SPEC>

Bit 10 - set this bit high to enable PWM generator1 timer equal a event generate

pub fn evt_op2_tea_en(&mut self) -> EVT_OP2_TEA_EN_W<'_, EVT_EN_SPEC>

Bit 11 - set this bit high to enable PWM generator2 timer equal a event generate

pub fn evt_op0_teb_en(&mut self) -> EVT_OP0_TEB_EN_W<'_, EVT_EN_SPEC>

Bit 12 - set this bit high to enable PWM generator0 timer equal b event generate

pub fn evt_op1_teb_en(&mut self) -> EVT_OP1_TEB_EN_W<'_, EVT_EN_SPEC>

Bit 13 - set this bit high to enable PWM generator1 timer equal b event generate

pub fn evt_op2_teb_en(&mut self) -> EVT_OP2_TEB_EN_W<'_, EVT_EN_SPEC>

Bit 14 - set this bit high to enable PWM generator2 timer equal b event generate

pub fn evt_f0_en(&mut self) -> EVT_F0_EN_W<'_, EVT_EN_SPEC>

Bit 15 - set this bit high to enable fault0 event generate

pub fn evt_f1_en(&mut self) -> EVT_F1_EN_W<'_, EVT_EN_SPEC>

Bit 16 - set this bit high to enable fault1 event generate

pub fn evt_f2_en(&mut self) -> EVT_F2_EN_W<'_, EVT_EN_SPEC>

Bit 17 - set this bit high to enable fault2 event generate

pub fn evt_f0_clr_en(&mut self) -> EVT_F0_CLR_EN_W<'_, EVT_EN_SPEC>

Bit 18 - set this bit high to enable fault0 clear event generate

pub fn evt_f1_clr_en(&mut self) -> EVT_F1_CLR_EN_W<'_, EVT_EN_SPEC>

Bit 19 - set this bit high to enable fault1 clear event generate

pub fn evt_f2_clr_en(&mut self) -> EVT_F2_CLR_EN_W<'_, EVT_EN_SPEC>

Bit 20 - set this bit high to enable fault2 clear event generate

pub fn evt_tz0_cbc_en(&mut self) -> EVT_TZ0_CBC_EN_W<'_, EVT_EN_SPEC>

Bit 21 - set this bit high to enable cycle by cycle trip0 event generate

pub fn evt_tz1_cbc_en(&mut self) -> EVT_TZ1_CBC_EN_W<'_, EVT_EN_SPEC>

Bit 22 - set this bit high to enable cycle by cycle trip1 event generate

pub fn evt_tz2_cbc_en(&mut self) -> EVT_TZ2_CBC_EN_W<'_, EVT_EN_SPEC>

Bit 23 - set this bit high to enable cycle by cycle trip2 event generate

pub fn evt_tz0_ost_en(&mut self) -> EVT_TZ0_OST_EN_W<'_, EVT_EN_SPEC>

Bit 24 - set this bit high to enable one shot trip0 event generate

pub fn evt_tz1_ost_en(&mut self) -> EVT_TZ1_OST_EN_W<'_, EVT_EN_SPEC>

Bit 25 - set this bit high to enable one shot trip1 event generate

pub fn evt_tz2_ost_en(&mut self) -> EVT_TZ2_OST_EN_W<'_, EVT_EN_SPEC>

Bit 26 - set this bit high to enable one shot trip2 event generate

pub fn evt_cap0_en(&mut self) -> EVT_CAP0_EN_W<'_, EVT_EN_SPEC>

Bit 27 - set this bit high to enable capture0 event generate

pub fn evt_cap1_en(&mut self) -> EVT_CAP1_EN_W<'_, EVT_EN_SPEC>

Bit 28 - set this bit high to enable capture1 event generate

pub fn evt_cap2_en(&mut self) -> EVT_CAP2_EN_W<'_, EVT_EN_SPEC>

Bit 29 - set this bit high to enable capture2 event generate

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual