Type Alias esp32h2::mcpwm0::task_en::R

pub type R = R<TASK_EN_SPEC>;

Register TASK_EN reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn task_cmpr0_a_up_en(&self) -> TASK_CMPR0_A_UP_EN_R

Bit 0 - set this bit high to enable PWM generator0 timer stamp A’s shadow register update task receive

pub fn task_cmpr1_a_up_en(&self) -> TASK_CMPR1_A_UP_EN_R

Bit 1 - set this bit high to enable PWM generator1 timer stamp A’s shadow register update task receive

pub fn task_cmpr2_a_up_en(&self) -> TASK_CMPR2_A_UP_EN_R

Bit 2 - set this bit high to enable PWM generator2 timer stamp A’s shadow register update task receive

pub fn task_cmpr0_b_up_en(&self) -> TASK_CMPR0_B_UP_EN_R

Bit 3 - set this bit high to enable PWM generator0 timer stamp B’s shadow register update task receive

pub fn task_cmpr1_b_up_en(&self) -> TASK_CMPR1_B_UP_EN_R

Bit 4 - set this bit high to enable PWM generator1 timer stamp B’s shadow register update task receive

pub fn task_cmpr2_b_up_en(&self) -> TASK_CMPR2_B_UP_EN_R

Bit 5 - set this bit high to enable PWM generator2 timer stamp B’s shadow register update task receive

pub fn task_gen_stop_en(&self) -> TASK_GEN_STOP_EN_R

Bit 6 - set this bit high to enable all PWM generate stop task receive

pub fn task_timer0_sync_en(&self) -> TASK_TIMER0_SYNC_EN_R

Bit 7 - set this bit high to enable timer0 sync task receive

pub fn task_timer1_sync_en(&self) -> TASK_TIMER1_SYNC_EN_R

Bit 8 - set this bit high to enable timer1 sync task receive

pub fn task_timer2_sync_en(&self) -> TASK_TIMER2_SYNC_EN_R

Bit 9 - set this bit high to enable timer2 sync task receive

pub fn task_timer0_period_up_en(&self) -> TASK_TIMER0_PERIOD_UP_EN_R

Bit 10 - set this bit high to enable timer0 period update task receive

pub fn task_timer1_period_up_en(&self) -> TASK_TIMER1_PERIOD_UP_EN_R

Bit 11 - set this bit high to enable timer1 period update task receive

pub fn task_timer2_period_up_en(&self) -> TASK_TIMER2_PERIOD_UP_EN_R

Bit 12 - set this bit high to enable timer2 period update task receive

pub fn task_tz0_ost_en(&self) -> TASK_TZ0_OST_EN_R

Bit 13 - set this bit high to enable one shot trip0 task receive

pub fn task_tz1_ost_en(&self) -> TASK_TZ1_OST_EN_R

Bit 14 - set this bit high to enable one shot trip1 task receive

pub fn task_tz2_ost_en(&self) -> TASK_TZ2_OST_EN_R

Bit 15 - set this bit high to enable one shot trip2 task receive

pub fn task_clr0_ost_en(&self) -> TASK_CLR0_OST_EN_R

Bit 16 - set this bit high to enable one shot trip0 clear task receive

pub fn task_clr1_ost_en(&self) -> TASK_CLR1_OST_EN_R

Bit 17 - set this bit high to enable one shot trip1 clear task receive

pub fn task_clr2_ost_en(&self) -> TASK_CLR2_OST_EN_R

Bit 18 - set this bit high to enable one shot trip2 clear task receive

pub fn task_cap0_en(&self) -> TASK_CAP0_EN_R

Bit 19 - set this bit high to enable capture0 task receive

pub fn task_cap1_en(&self) -> TASK_CAP1_EN_R

Bit 20 - set this bit high to enable capture1 task receive

pub fn task_cap2_en(&self) -> TASK_CAP2_EN_R

Bit 21 - set this bit high to enable capture2 task receive