Module esp32h2::pcnt::u_conf0

Configuration register 0 for unit %s

Structs

U_CONF0_SPEC

Configuration register 0 for unit %s

Type Aliases

CH0_HCTRL_MODE_R

Field CH0_HCTRL_MODE reader - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is high. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH0_HCTRL_MODE_W

Field CH0_HCTRL_MODE writer - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is high. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH0_LCTRL_MODE_R

Field CH0_LCTRL_MODE reader - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is low. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH0_LCTRL_MODE_W

Field CH0_LCTRL_MODE writer - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is low. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH0_NEG_MODE_R

Field CH0_NEG_MODE reader - This register sets the behavior when the signal input of channel 0 detects a negative edge. 1: Increase the counter.2: Decrease the counter.0, 3: No effect on counter

CH0_NEG_MODE_W

Field CH0_NEG_MODE writer - This register sets the behavior when the signal input of channel 0 detects a negative edge. 1: Increase the counter.2: Decrease the counter.0, 3: No effect on counter

CH0_POS_MODE_R

Field CH0_POS_MODE reader - This register sets the behavior when the signal input of channel 0 detects a positive edge. 1: Increase the counter.2: Decrease the counter.0, 3: No effect on counter

CH0_POS_MODE_W

Field CH0_POS_MODE writer - This register sets the behavior when the signal input of channel 0 detects a positive edge. 1: Increase the counter.2: Decrease the counter.0, 3: No effect on counter

CH1_HCTRL_MODE_R

Field CH1_HCTRL_MODE reader - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is high. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH1_HCTRL_MODE_W

Field CH1_HCTRL_MODE writer - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is high. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH1_LCTRL_MODE_R

Field CH1_LCTRL_MODE reader - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is low. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH1_LCTRL_MODE_W

Field CH1_LCTRL_MODE writer - This register configures how the CH%s_POS_MODE/CH%s_NEG_MODE settings will be modified when the control signal is low. 0: No modification.1: Invert behavior (increase -> decrease, decrease -> increase).2, 3: Inhibit counter modification

CH1_NEG_MODE_R

Field CH1_NEG_MODE reader - This register sets the behavior when the signal input of channel 1 detects a negative edge. 1: Increment the counter.2: Decrement the counter.0, 3: No effect on counter

CH1_NEG_MODE_W

Field CH1_NEG_MODE writer - This register sets the behavior when the signal input of channel 1 detects a negative edge. 1: Increment the counter.2: Decrement the counter.0, 3: No effect on counter

CH1_POS_MODE_R

Field CH1_POS_MODE reader - This register sets the behavior when the signal input of channel 1 detects a positive edge. 1: Increment the counter.2: Decrement the counter.0, 3: No effect on counter

CH1_POS_MODE_W

Field CH1_POS_MODE writer - This register sets the behavior when the signal input of channel 1 detects a positive edge. 1: Increment the counter.2: Decrement the counter.0, 3: No effect on counter

FILTER_EN_R

Field FILTER_EN reader - This is the enable bit for unit %s’s input filter.

FILTER_EN_W

Field FILTER_EN writer - This is the enable bit for unit %s’s input filter.

FILTER_THRES_R

Field FILTER_THRES reader - This sets the maximum threshold, in APB_CLK cycles, for the filter. Any pulses with width less than this will be ignored when the filter is enabled.

FILTER_THRES_W

Field FILTER_THRES writer - This sets the maximum threshold, in APB_CLK cycles, for the filter. Any pulses with width less than this will be ignored when the filter is enabled.

R

Register U%s_CONF0 reader

THR_H_LIM_EN_R

Field THR_H_LIM_EN reader - This is the enable bit for unit %s’s thr_h_lim comparator. Configures it to enable the high limit interrupt.

THR_H_LIM_EN_W

Field THR_H_LIM_EN writer - This is the enable bit for unit %s’s thr_h_lim comparator. Configures it to enable the high limit interrupt.

THR_L_LIM_EN_R

Field THR_L_LIM_EN reader - This is the enable bit for unit %s’s thr_l_lim comparator. Configures it to enable the low limit interrupt.

THR_L_LIM_EN_W

Field THR_L_LIM_EN writer - This is the enable bit for unit %s’s thr_l_lim comparator. Configures it to enable the low limit interrupt.

THR_THRES0_EN_R

Field THR_THRES0_EN reader - This is the enable bit for unit %s’s thres0 comparator.

THR_THRES0_EN_W

Field THR_THRES0_EN writer - This is the enable bit for unit %s’s thres0 comparator.

THR_THRES1_EN_R

Field THR_THRES1_EN reader - This is the enable bit for unit %s’s thres1 comparator.

THR_THRES1_EN_W

Field THR_THRES1_EN writer - This is the enable bit for unit %s’s thres1 comparator.

THR_ZERO_EN_R

Field THR_ZERO_EN reader - This is the enable bit for unit %s’s zero comparator.

THR_ZERO_EN_W

Field THR_ZERO_EN writer - This is the enable bit for unit %s’s zero comparator.

W

Register U%s_CONF0 writer