Module esp32h2::pcr::parl_clk_rx_conf

PARL_CLK_RX configuration register

Structs

PARL_CLK_RX_CONF_SPEC

PARL_CLK_RX configuration register

Type Aliases

PARL_CLK_RX_DIV_NUM_R

Field PARL_CLK_RX_DIV_NUM reader - The integral part of the frequency divider factor of the parl rx clock.

PARL_CLK_RX_DIV_NUM_W

Field PARL_CLK_RX_DIV_NUM writer - The integral part of the frequency divider factor of the parl rx clock.

PARL_CLK_RX_EN_R

Field PARL_CLK_RX_EN reader - Set 1 to enable parl rx clock

PARL_CLK_RX_EN_W

Field PARL_CLK_RX_EN writer - Set 1 to enable parl rx clock

PARL_CLK_RX_SEL_R

Field PARL_CLK_RX_SEL reader - set this field to select clock-source. 0(default): XTAL, 1: 240MHz, 2: FOSC, 3: user clock from pad.

PARL_CLK_RX_SEL_W

Field PARL_CLK_RX_SEL writer - set this field to select clock-source. 0(default): XTAL, 1: 240MHz, 2: FOSC, 3: user clock from pad.

PARL_RX_RST_EN_R

Field PARL_RX_RST_EN reader - Set 0 to reset parl rx module

PARL_RX_RST_EN_W

Field PARL_RX_RST_EN writer - Set 0 to reset parl rx module

R

Register PARL_CLK_RX_CONF reader

W

Register PARL_CLK_RX_CONF writer