Module esp32h2::spi0::spi_mem_ctrl

SPI0 control register.

Structs

SPI_MEM_CTRL_SPEC

SPI0 control register.

Type Aliases

R

Register SPI_MEM_CTRL reader

SPI_MEM_DATA_IE_ALWAYS_ON_R

Field SPI_MEM_DATA_IE_ALWAYS_ON reader - When accesses to flash, 1: the IE signals of pads connected to SPI_IO[7:0] are always 1. 0: Others.

SPI_MEM_DATA_IE_ALWAYS_ON_W

Field SPI_MEM_DATA_IE_ALWAYS_ON writer - When accesses to flash, 1: the IE signals of pads connected to SPI_IO[7:0] are always 1. 0: Others.

SPI_MEM_DQS_IE_ALWAYS_ON_R

Field SPI_MEM_DQS_IE_ALWAYS_ON reader - When accesses to flash, 1: the IE signals of pads connected to SPI_DQS are always 1. 0: Others.

SPI_MEM_D_POL_R

Field SPI_MEM_D_POL reader - The bit is used to set MOSI line polarity, 1: high 0, low

SPI_MEM_D_POL_W

Field SPI_MEM_D_POL writer - The bit is used to set MOSI line polarity, 1: high 0, low

SPI_MEM_FADDR_OCT_R

Field SPI_MEM_FADDR_OCT reader - Apply 8 signals during address phase 1:enable 0: disable

SPI_MEM_FASTRD_MODE_R

Field SPI_MEM_FASTRD_MODE reader - This bit enable the bits: SPI_MEM_FREAD_QIO, SPI_MEM_FREAD_DIO, SPI_MEM_FREAD_QOUT and SPI_MEM_FREAD_DOUT. 1: enable 0: disable.

SPI_MEM_FASTRD_MODE_W

Field SPI_MEM_FASTRD_MODE writer - This bit enable the bits: SPI_MEM_FREAD_QIO, SPI_MEM_FREAD_DIO, SPI_MEM_FREAD_QOUT and SPI_MEM_FREAD_DOUT. 1: enable 0: disable.

SPI_MEM_FCMD_OCT_R

Field SPI_MEM_FCMD_OCT reader - Apply 8 signals during command phase 1:enable 0: disable

SPI_MEM_FCMD_QUAD_R

Field SPI_MEM_FCMD_QUAD reader - Apply 4 signals during command phase 1:enable 0: disable

SPI_MEM_FCMD_QUAD_W

Field SPI_MEM_FCMD_QUAD writer - Apply 4 signals during command phase 1:enable 0: disable

SPI_MEM_FDIN_OCT_R

Field SPI_MEM_FDIN_OCT reader - Apply 8 signals during read-data phase 1:enable 0: disable

SPI_MEM_FDOUT_OCT_R

Field SPI_MEM_FDOUT_OCT reader - Apply 8 signals during write-data phase 1:enable 0: disable

SPI_MEM_FDUMMY_RIN_R

Field SPI_MEM_FDUMMY_RIN reader - In an MSPI read data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller in the first half part of dummy phase. It is used to mask invalid SPI_DQS in the half part of dummy phase.

SPI_MEM_FDUMMY_RIN_W

Field SPI_MEM_FDUMMY_RIN writer - In an MSPI read data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller in the first half part of dummy phase. It is used to mask invalid SPI_DQS in the half part of dummy phase.

SPI_MEM_FDUMMY_WOUT_R

Field SPI_MEM_FDUMMY_WOUT reader - In an MSPI write data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller in the second half part of dummy phase. It is used to pre-drive flash.

SPI_MEM_FDUMMY_WOUT_W

Field SPI_MEM_FDUMMY_WOUT writer - In an MSPI write data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller in the second half part of dummy phase. It is used to pre-drive flash.

SPI_MEM_FREAD_DIO_R

Field SPI_MEM_FREAD_DIO reader - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

SPI_MEM_FREAD_DIO_W

Field SPI_MEM_FREAD_DIO writer - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

SPI_MEM_FREAD_DUAL_R

Field SPI_MEM_FREAD_DUAL reader - In the read operations, read-data phase apply 2 signals. 1: enable 0: disable.

SPI_MEM_FREAD_DUAL_W

Field SPI_MEM_FREAD_DUAL writer - In the read operations, read-data phase apply 2 signals. 1: enable 0: disable.

SPI_MEM_FREAD_QIO_R

Field SPI_MEM_FREAD_QIO reader - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

SPI_MEM_FREAD_QIO_W

Field SPI_MEM_FREAD_QIO writer - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

SPI_MEM_FREAD_QUAD_R

Field SPI_MEM_FREAD_QUAD reader - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

SPI_MEM_FREAD_QUAD_W

Field SPI_MEM_FREAD_QUAD writer - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

SPI_MEM_Q_POL_R

Field SPI_MEM_Q_POL reader - The bit is used to set MISO line polarity, 1: high 0, low

SPI_MEM_Q_POL_W

Field SPI_MEM_Q_POL writer - The bit is used to set MISO line polarity, 1: high 0, low

SPI_MEM_WDUMMY_ALWAYS_OUT_R

Field SPI_MEM_WDUMMY_ALWAYS_OUT reader - In the dummy phase of an MSPI write data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller.

SPI_MEM_WDUMMY_ALWAYS_OUT_W

Field SPI_MEM_WDUMMY_ALWAYS_OUT writer - In the dummy phase of an MSPI write data transfer when accesses to flash, the level of SPI_IO[7:0] is output by the MSPI controller.

SPI_MEM_WDUMMY_DQS_ALWAYS_OUT_R

Field SPI_MEM_WDUMMY_DQS_ALWAYS_OUT reader - In the dummy phase of an MSPI write data transfer when accesses to flash, the level of SPI_DQS is output by the MSPI controller.

SPI_MEM_WP_R

Field SPI_MEM_WP reader - Write protect signal output when SPI is idle. 1: output high, 0: output low.

SPI_MEM_WP_W

Field SPI_MEM_WP writer - Write protect signal output when SPI is idle. 1: output high, 0: output low.

W

Register SPI_MEM_CTRL writer