Module esp32h2::spi0::spi_mem_dpa_ctrl

SPI memory cryption DPA register

Structs

SPI_MEM_DPA_CTRL_SPEC

SPI memory cryption DPA register

Type Aliases

R

Register SPI_MEM_DPA_CTRL reader

SPI_CRYPT_CALC_D_DPA_EN_R

Field SPI_CRYPT_CALC_D_DPA_EN reader - Only available when SPI_CRYPT_SECURITY_LEVEL is not 0. 1: Enable DPA in the calculation that using key 1 or key 2. 0: Enable DPA only in the calculation that using key 1.

SPI_CRYPT_CALC_D_DPA_EN_W

Field SPI_CRYPT_CALC_D_DPA_EN writer - Only available when SPI_CRYPT_SECURITY_LEVEL is not 0. 1: Enable DPA in the calculation that using key 1 or key 2. 0: Enable DPA only in the calculation that using key 1.

SPI_CRYPT_DPA_SELECT_REGISTER_R

Field SPI_CRYPT_DPA_SELECT_REGISTER reader - 1: MSPI XTS DPA clock gate is controlled by SPI_CRYPT_CALC_D_DPA_EN and SPI_CRYPT_SECURITY_LEVEL. 0: Controlled by efuse bits.

SPI_CRYPT_DPA_SELECT_REGISTER_W

Field SPI_CRYPT_DPA_SELECT_REGISTER writer - 1: MSPI XTS DPA clock gate is controlled by SPI_CRYPT_CALC_D_DPA_EN and SPI_CRYPT_SECURITY_LEVEL. 0: Controlled by efuse bits.

SPI_CRYPT_SECURITY_LEVEL_R

Field SPI_CRYPT_SECURITY_LEVEL reader - Set the security level of spi mem cryption. 0: Shut off cryption DPA funtion. 1-7: The bigger the number is, the more secure the cryption is. (Note that the performance of cryption will decrease together with this number increasing)

SPI_CRYPT_SECURITY_LEVEL_W

Field SPI_CRYPT_SECURITY_LEVEL writer - Set the security level of spi mem cryption. 0: Shut off cryption DPA funtion. 1-7: The bigger the number is, the more secure the cryption is. (Note that the performance of cryption will decrease together with this number increasing)

W

Register SPI_MEM_DPA_CTRL writer