Module esp32h2::twai0::status

TWAI status register.

Structs

STATUS_SPEC

TWAI status register.

Type Aliases

ERR_R

Field ERR reader - 1: error, at least one of the error counters has reached or exceeded the CPU warning limit defined by the Error Warning Limit Register (EWLR). 0: ok, both error counters are below the warning limit

MISS_R

Field MISS reader - 1: current message is destroyed because of FIFO overflow.

NODE_BUS_OFF_R

Field NODE_BUS_OFF reader - 1: bus-off, the TWAI controller is not involved in bus activities. 0: bus-on, the TWAI controller is involved in bus activities

OVERRUN_R

Field OVERRUN reader - 1: overrun, a message was lost because there was not enough space for that message in the RXFIFO. 0: absent, no data overrun has occurred since the last clear data overrun command was given

R

Register STATUS reader

RECEIVE_BUFFER_R

Field RECEIVE_BUFFER reader - 1: full, one or more complete messages are available in the RXFIFO. 0: empty, no message is available

RECEIVE_R

Field RECEIVE reader - 1: receive, the TWAI controller is receiving a message. 0: idle

TRANSMISSION_COMPLETE_R

Field TRANSMISSION_COMPLETE reader - 1: complete, last requested transmission has been successfully completed. 0: incomplete, previously requested transmission is not yet completed

TRANSMIT_BUFFER_R

Field TRANSMIT_BUFFER reader - 1: released, the CPU may write a message into the transmit buffer. 0: locked, the CPU cannot access the transmit buffer, a message is either waiting for transmission or is in the process of being transmitted

TRANSMIT_R

Field TRANSMIT reader - 1: transmit, the TWAI controller is transmitting a message. 0: idle