Module esp32h2::uart0::conf0

a

Structs

CONF0_SPEC

a

Type Aliases

AUTOBAUD_EN_R

Field AUTOBAUD_EN reader - This is the enable bit for detecting baudrate.

AUTOBAUD_EN_W

Field AUTOBAUD_EN writer - This is the enable bit for detecting baudrate.

BIT_NUM_R

Field BIT_NUM reader - This register is used to set the length of data.

BIT_NUM_W

Field BIT_NUM writer - This register is used to set the length of data.

DIS_RX_DAT_OVF_R

Field DIS_RX_DAT_OVF reader - Disable UART Rx data overflow detect.

DIS_RX_DAT_OVF_W

Field DIS_RX_DAT_OVF writer - Disable UART Rx data overflow detect.

ERR_WR_MASK_R

Field ERR_WR_MASK reader - 1’h1: Receiver stops storing data into FIFO when data is wrong. 1’h0: Receiver stores the data even if the received data is wrong.

ERR_WR_MASK_W

Field ERR_WR_MASK writer - 1’h1: Receiver stops storing data into FIFO when data is wrong. 1’h0: Receiver stores the data even if the received data is wrong.

IRDA_DPLX_R

Field IRDA_DPLX reader - Set this bit to enable IrDA loopback mode.

IRDA_DPLX_W

Field IRDA_DPLX writer - Set this bit to enable IrDA loopback mode.

IRDA_EN_R

Field IRDA_EN reader - Set this bit to enable IrDA protocol.

IRDA_EN_W

Field IRDA_EN writer - Set this bit to enable IrDA protocol.

IRDA_RX_INV_R

Field IRDA_RX_INV reader - Set this bit to invert the level of IrDA receiver.

IRDA_RX_INV_W

Field IRDA_RX_INV writer - Set this bit to invert the level of IrDA receiver.

IRDA_TX_EN_R

Field IRDA_TX_EN reader - This is the start enable bit for IrDA transmitter.

IRDA_TX_EN_W

Field IRDA_TX_EN writer - This is the start enable bit for IrDA transmitter.

IRDA_TX_INV_R

Field IRDA_TX_INV reader - Set this bit to invert the level of IrDA transmitter.

IRDA_TX_INV_W

Field IRDA_TX_INV writer - Set this bit to invert the level of IrDA transmitter.

IRDA_WCTL_R

Field IRDA_WCTL reader - 1’h1: The IrDA transmitter’s 11th bit is the same as 10th bit. 1’h0: Set IrDA transmitter’s 11th bit to 0.

IRDA_WCTL_W

Field IRDA_WCTL writer - 1’h1: The IrDA transmitter’s 11th bit is the same as 10th bit. 1’h0: Set IrDA transmitter’s 11th bit to 0.

LOOPBACK_R

Field LOOPBACK reader - Set this bit to enable uart loopback test mode.

LOOPBACK_W

Field LOOPBACK writer - Set this bit to enable uart loopback test mode.

MEM_CLK_EN_R

Field MEM_CLK_EN reader - UART memory clock gate enable signal.

MEM_CLK_EN_W

Field MEM_CLK_EN writer - UART memory clock gate enable signal.

PARITY_EN_R

Field PARITY_EN reader - Set this bit to enable uart parity check.

PARITY_EN_W

Field PARITY_EN writer - Set this bit to enable uart parity check.

PARITY_R

Field PARITY reader - This register is used to configure the parity check mode.

PARITY_W

Field PARITY writer - This register is used to configure the parity check mode.

R

Register CONF0 reader

RXD_INV_R

Field RXD_INV reader - Set this bit to inverse the level value of uart rxd signal.

RXD_INV_W

Field RXD_INV writer - Set this bit to inverse the level value of uart rxd signal.

RXFIFO_RST_R

Field RXFIFO_RST reader - Set this bit to reset the uart receive-FIFO.

RXFIFO_RST_W

Field RXFIFO_RST writer - Set this bit to reset the uart receive-FIFO.

STOP_BIT_NUM_R

Field STOP_BIT_NUM reader - This register is used to set the length of stop bit.

STOP_BIT_NUM_W

Field STOP_BIT_NUM writer - This register is used to set the length of stop bit.

SW_RTS_R

Field SW_RTS reader - This register is used to configure the software rts signal which is used in software flow control.

SW_RTS_W

Field SW_RTS writer - This register is used to configure the software rts signal which is used in software flow control.

TXD_BRK_R

Field TXD_BRK reader - Set this bit to enbale transmitter to send NULL when the process of sending data is done.

TXD_BRK_W

Field TXD_BRK writer - Set this bit to enbale transmitter to send NULL when the process of sending data is done.

TXD_INV_R

Field TXD_INV reader - Set this bit to inverse the level value of uart txd signal.

TXD_INV_W

Field TXD_INV writer - Set this bit to inverse the level value of uart txd signal.

TXFIFO_RST_R

Field TXFIFO_RST reader - Set this bit to reset the uart transmit-FIFO.

TXFIFO_RST_W

Field TXFIFO_RST writer - Set this bit to reset the uart transmit-FIFO.

TX_FLOW_EN_R

Field TX_FLOW_EN reader - Set this bit to enable flow control function for transmitter.

TX_FLOW_EN_W

Field TX_FLOW_EN writer - Set this bit to enable flow control function for transmitter.

W

Register CONF0 writer